not quite capturing sales data

+++ app/src/main/java/com/example/narddogsinventory/SalesCategory.kt
@@ -5,6 +5,6 @@ data class SalesCategory(
     var totItems : Int? = null,
     var totDollars : Float? = null,
     var margin : Float? = null,
-    var averageSale : Int? = null,
+    var averageSale : Float? = null,
     var totCost : Float? = null
 )
